Removed the parameter num from method writeOut of classes Data and added method/costructor to set the sensor in the class RGBDData. For the moment the lines: 94 --> int num = this->_rgbdCameraSensor->getNum(); 84 --> os << _rgbdCameraSensor->getParameter()->id();
in the methods writeOut and write are commented since they cause a segmentation fault (strange since the sensor, when the data
is created inside the relative sensorHandler, is setted).

Commenting the line 94 cause the writeOut to save all the images on the same file and therefore the output will be only one image
since each new image overwrite the old one.